requirements.txt
nordigen==1.4.0
python-dotenv==1.0.1
Requests==2.31.0
